+++ src/components/calcite-block/calcite-block.tsx
@@ -140,7 +140,7 @@ export class CalciteBlock {
     ) : !disabled ? (
       <slot name={SLOTS.control} />
     ) : null;
-    const hasIcon = this.el.querySelector("[slot=icon]");
+    const hasIcon = el.querySelector(`[slot=${SLOTS.icon}]`);
     const headerContent = (
       <header class={CSS.header}>
         {hasIcon ? (
@@ -156,14 +156,8 @@ export class CalciteBlock {
       </header>
     );
 
-    let hasControl = false;
-    let hasContent = false;
-
-    for (let i = 0; i < el.children.length; i++) {
-      const isControl = el.children[i].slot === SLOTS.control;
-      hasControl = hasControl || isControl;
-      hasContent = hasContent || !isControl;
-    }
+    const controlSlot = el.querySelector(`[slot=${SLOTS.control}]`);
+    const hasContent = el.children.length > (controlSlot ? 1 : 0);
 
     const headerNode = (
       <div class={CSS.headerContainer}>
@@ -175,7 +169,7 @@ export class CalciteBlock {
             title={toggleLabel}
           >
             {headerContent}
-            {hasControl ? null : (
+            {controlSlot ? null : (
               <CalciteIcon
                 size="16"
                 path={open ? chevronUp16 : chevronDown16}
@@ -194,8 +188,9 @@ export class CalciteBlock {
         <article aria-expanded={collapsible ? (open ? "true" : "false") : null} aria-busy={loading}>
           {headerNode}
           <div class={CSS.content} hidden={!hasContent || !open}>
-            <slot />
-            {loading || disabled ? <CalciteScrim loading={false}></CalciteScrim> : null}
+            <CalciteScrim loading={loading} disabled={disabled}>
+              <slot />
+            </CalciteScrim>
           </div>
         </article>
       </Host>