+++ src/components/calcite-pick-list-item/calcite-pick-list-item.e2e.ts
@@ -1,4 +1,5 @@
 import { newE2EPage } from "@stencil/core/testing";
+import { accessible } from "../../tests/commonTests";
 
 describe("calcite-pick-list-item", () => {
   it("should render", async () => {
@@ -11,6 +12,9 @@ describe("calcite-pick-list-item", () => {
     expect(isVisible).toBe(true);
   });
 
+  it("is accessible", async () =>
+    accessible(`<calcite-pick-list-item text-label="Sample" value="one"></calcite-pick-list-item>`));
+
   it("should toggle selected attribute when clicked", async () => {
     const page = await newE2EPage();
 
@@ -21,5 +25,30 @@ describe("calcite-pick-list-item", () => {
     expect(await item.getProperty("selected")).toBe(true);
     await item.click();
     expect(await item.getProperty("selected")).toBe(false);
+  });
+
+  it("should fire event calciteListItemChange when item is clicked", async () => {
+    const page = await newE2EPage();
+
+    await page.setContent(`<calcite-pick-list-item text-label="test" value="example"></calcite-pick-list-item>`);
+    const item = await page.find("calcite-pick-list-item");
+    await page.evaluate(() => {
+      document.addEventListener("calciteListItemChange", (event: CustomEvent) => {
+        (window as any).eventDetail = event.detail;
+      });
+    });
+
+    await item.click();
+
+    await page.waitForChanges();
+
+    const eventDetail: any = await page.evaluateHandle(() => {
+      return (window as any).eventDetail;
+    });
+    const properties = await eventDetail.getProperties();
+    expect(properties.get("item")).toBeDefined();
+    expect(properties.get("value")._remoteObject.value).toBe("example");
+    expect(properties.get("selected")._remoteObject.value).toBe(true);
+    expect(properties.get("shiftPressed")._remoteObject.value).toBe(false);
   });
 });