+++ src/components/calcite-pick-list/calcite-pick-list.tsx
@@ -27,6 +27,7 @@ const {
 } = sharedListMethods;
 
 /**
+ * @slot - A slot for adding pick-list-item elements or pick-list-groups elements. Items are displayed as a vertical list.
  * @slot menu-actions - A slot for adding a button + menu combo for performing actions like sorting.
  */
 @Component({
@@ -101,6 +102,8 @@ export class CalcitePickList {
   observer = new MutationObserver(mutationObserverCallback.bind(this));
 
   @Element() el: HTMLCalcitePickListElement;
+
+  emitCalciteListChange: () => void;
 
   // --------------------------------------------------------------------------
   //