+++ src/components/calcite-pick-list/shared-list-render.tsx
@@ -1,9 +1,10 @@
 import { Host, h } from "@stencil/core";
 import CalciteScrim from "../utils/CalciteScrim";
 import { VNode } from "@esri/calcite-components/dist/types/stencil.core";
+import { CSS } from "./resources";
 
 const renderScrim = (loading, disabled): VNode => {
-  return loading || disabled ? <CalciteScrim loading={loading}></CalciteScrim> : null;
+  return <CalciteScrim loading={loading} disabled={disabled} />;
 };
 
 export const List = ({ props, text, ...rest }): VNode => {
@@ -11,7 +12,7 @@ export const List = ({ props, text, ...rest }): VNode => {
   const { filterPlaceholder } = text;
   return (
     <Host aria-disabled={disabled} aria-busy={loading} {...rest}>
-      <header>
+      <header class={{ [CSS.sticky]: true }}>
         {filterEnabled ? (
           <calcite-filter
             data={dataForFilter}