+++ src/components/calcite-value-list-item/calcite-value-list-item.tsx
@@ -1,6 +1,13 @@
 import { Component, Element, Host, Method, Prop, h } from "@stencil/core";
 import { ICON_TYPES } from "../calcite-pick-list/resources";
+import guid from "../utils/guid";
+import { CSS } from "../calcite-pick-list-item/resources";
+import { drag16 } from "@esri/calcite-ui-icons";
+import CalciteIcon from "../utils/CalciteIcon";
 
+/**
+ * @slot secondaryAction - A slot intended for adding a calcite-action or calcite-button. Placed at the end of the item.
+ */
 @Component({
   tag: "calcite-value-list-item",
   styleUrl: "./calcite-value-list-item.scss",
@@ -16,32 +23,37 @@ export class CalciteValueListItem {
   /**
    * Compact reduces the size of the item.
    */
-  @Prop({ reflect: true }) compact = false;
+  @Prop({ reflect: true }) compact? = false;
 
   /**
    * When true, the item cannot be clicked and is visually muted
    */
-  @Prop({ reflect: true }) disabled = false;
+  @Prop({ reflect: true }) disabled? = false;
+
+  /**
+   * @internal - stores the activated state of the drag handle.
+   */
+  @Prop({ mutable: true }) handleActivated? = false;
 
   /**
    * Determines the icon SVG symbol that will be shown. Options are circle, square, grid or null.
    */
-  @Prop({ reflect: true }) icon: ICON_TYPES | null = null;
+  @Prop({ reflect: true }) icon?: ICON_TYPES | null = null;
 
   /**
    * Used to provide additional metadata to an item, primarily used when the parent list has a filter.
    */
-  @Prop() metadata: object;
+  @Prop() metadata?: object;
 
   /**
    * Set this to true to pre-select an item. Toggles when an item is checked/unchecked.
    */
-  @Prop() selected = false;
+  @Prop({ reflect: true, mutable: true }) selected? = false;
 
   /**
    * The main label for this item. Appears next to the icon.
    */
-  @Prop({ reflect: true }) textLabel: string;
+  @Prop({ reflect: true }) textLabel!: string;
 
   /**
    * An optional description for this item. Will appear below the label text.
@@ -51,7 +63,7 @@ export class CalciteValueListItem {
   /**
    * A unique value used to identify this item - similar to the value attribute on an <input>.
    */
-  @Prop({ reflect: true }) value: string;
+  @Prop({ reflect: true }) value!: string;
 
   // --------------------------------------------------------------------------
   //
@@ -63,14 +75,16 @@ export class CalciteValueListItem {
 
   pickListItem: HTMLCalcitePickListItemElement = null;
 
+  guid = `calcite-value-list-item-${guid()}`;
+
   // --------------------------------------------------------------------------
   //
   //  Public Methods
   //
   // --------------------------------------------------------------------------
 
-  @Method() async toggleSelected(coerce?: boolean, emit = false) {
-    this.pickListItem.toggleSelected(coerce, emit);
+  @Method() async toggleSelected(coerce?: boolean) {
+    this.pickListItem.toggleSelected(coerce);
   }
 
   // --------------------------------------------------------------------------
@@ -81,22 +95,51 @@ export class CalciteValueListItem {
 
   getPickListRef = (el) => (this.pickListItem = el as HTMLCalcitePickListItemElement);
 
+  handleKeyDown = (event: KeyboardEvent) => {
+    if (event.key === " ") {
+      this.handleActivated = !this.handleActivated;
+    }
+  };
+
+  handleBlur = () => {
+    this.handleActivated = false;
+  };
+
   // --------------------------------------------------------------------------
   //
   //  Render Methods
   //
   // --------------------------------------------------------------------------
 
+  renderHandle() {
+    const { icon } = this;
+    if (icon === ICON_TYPES.grip) {
+      return (
+        <span
+          role="button"
+          class={{ [CSS.handle]: true, [CSS.handleActivated]: this.handleActivated }}
+          tabindex="0"
+          data-js-handle="true"
+          aria-pressed={this.handleActivated}
+          onKeyDown={this.handleKeyDown}
+          onBlur={this.handleBlur}
+        >
+          <CalciteIcon size="16" path={drag16} />
+        </span>
+      );
+    }
+  }
+
   render() {
     return (
-      <Host>
+      <Host data-id={this.guid}>
+        {this.renderHandle()}
         <calcite-pick-list-item
           compact={this.compact}
           ref={this.getPickListRef}
           disabled={this.disabled}
           selected={this.selected}
           metadata={this.metadata}
-          icon={this.icon}
           textLabel={this.textLabel}
           textDescription={this.textDescription}
           value={this.value}