+++ src/components/calcite-value-list/calcite-value-list.tsx
@@ -29,6 +29,7 @@ const {
 } = sharedListMethods;
 
 /**
+ * @slot - A slot for adding pick-list-item elements or pick-list-groups elements. Items are displayed as a vertical list.
  * @slot menu-actions - A slot for adding a button + menu combo for performing actions like sorting.
  */
 @Component({
@@ -96,13 +97,9 @@ export class CalciteValueList {
 
   sortables: Sortable[] = [];
 
-  // --------------------------------------------------------------------------
-  //
-  //  Private Properties
-  //
-  // --------------------------------------------------------------------------
-
   @Element() el: HTMLCalciteValueListItemElement;
+
+  emitCalciteListChange: () => void;
 
   // --------------------------------------------------------------------------
   //
@@ -202,6 +199,56 @@ export class CalciteValueList {
 
   getItemData = getItemData.bind(this);
 
+  keyDownHandler = (event) => {
+    const handleElement = event.composedPath().find((item) => {
+      return item.dataset?.jsHandle;
+    });
+    const valueListElement = event.composedPath().find((item) => {
+      return item.tagName?.toLowerCase() === "calcite-value-list-item";
+    });
+    // Only trigger keyboard sorting when the internal drag handle is focused and activated
+    if (!handleElement || !valueListElement.handleActivated) {
+      return;
+    }
+    const lastIndex = this.items.length - 1;
+    const value = valueListElement.value;
+    const startingIndex = this.items.findIndex((item) => {
+      return item.value === value;
+    });
+    let appendInstead = false;
+    let buddyIndex;
+    switch (event.key) {
+      case "ArrowUp":
+        event.preventDefault();
+        if (startingIndex === 0) {
+          appendInstead = true;
+        } else {
+          buddyIndex = startingIndex - 1;
+        }
+        break;
+      case "ArrowDown":
+        event.preventDefault();
+        if (startingIndex === lastIndex) {
+          buddyIndex = 0;
+        } else if (startingIndex === lastIndex - 1) {
+          appendInstead = true;
+        } else {
+          buddyIndex = startingIndex + 2;
+        }
+        break;
+      default:
+        return;
+    }
+    if (appendInstead) {
+      valueListElement.parentElement.appendChild(valueListElement);
+    } else {
+      valueListElement.parentElement.insertBefore(valueListElement, this.items[buddyIndex]);
+    }
+
+    handleElement.focus();
+    valueListElement.handleActivated = true;
+  };
+
   // --------------------------------------------------------------------------
   //
   //  Public Methods
@@ -227,6 +274,6 @@ export class CalciteValueList {
   }
 
   render() {
-    return <List props={this} text={TEXT} />;
+    return <List props={this} text={TEXT} onKeyDown={this.keyDownHandler} />;
   }
 }